리팩토링
데이터 fixture 생성
