+++ __fixtures__/restaurants.js
@@ -1,0 +1,11 @@
+const restaurants = [
+  {
+    id: 3,
+    categoryId: 4,
+    name: '마법사주방',
+    address: '서울 강남구 강남대로94길 9',
+    information: '마법사주방 in 서울 강남구 강남대로94길 9',
+  },
+];
+
+export default restaurants;
