+++ src/actions.test.js
@@ -16,6 +16,10 @@ import {
   setRestaurants,
 } from './actions';
 
+import regions from '../__fixtures__/regions';
+import categories from '../__fixtures__/categories';
+import restaurants from '../__fixtures__/restaurants';
+
 jest.mock('./services/api');
 
 const middlewares = [thunk];
@@ -23,16 +27,6 @@ const mockStore = configureStore(middlewares);
 
 describe('actions', () => {
   describe('loadRestaurantRegions', () => {
-    const regions = [
-      { id: 1, name: '서울' },
-      { id: 2, name: '대전' },
-      { id: 3, name: '대구' },
-      { id: 4, name: '부산' },
-      { id: 5, name: '광주' },
-      { id: 6, name: '강원도' },
-      { id: 7, name: '인천' },
-    ];
-
     fetchRestaurantRegion.mockResolvedValue(regions);
 
     const store = mockStore({});
@@ -49,16 +43,6 @@ describe('actions', () => {
   });
 
   describe('loadRestaurantCategories', () => {
-    const categories = [
-      { id: 1, name: '한식' },
-      { id: 2, name: '중식' },
-      { id: 3, name: '일식' },
-      { id: 4, name: '양식' },
-      { id: 5, name: '분식' },
-      { id: 6, name: '과자' },
-      { id: 7, name: '치킨' },
-    ];
-
     fetchRestaurantCategories.mockResolvedValue(categories);
 
     const store = mockStore({});
@@ -75,16 +59,6 @@ describe('actions', () => {
   });
 
   describe('loadRestaurnats', () => {
-    const restaurants = [
-      {
-        id: 3,
-        categoryId: 4,
-        name: '마법사주방',
-        address: '서울 강남구 강남대로94길 9',
-        information: '마법사주방 in 서울 강남구 강남대로94길 9',
-      },
-    ];
-
     fetchRestaurants.mockResolvedValue(restaurants);
 
     context('with selected restaurant region and category', () => {
